# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LyoDex API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /storage/blog-cover-upload:
    post:
      operationId: uploadBlogCoverImage
      tags: [Storage]
      summary: Upload a blog cover image (admin only). Send raw image bytes with correct Content-Type header.
      requestBody:
        required: true
        content:
          image/jpeg:
            schema:
              type: string
              format: binary
          image/png:
            schema:
              type: string
              format: binary
          image/webp:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Upload succeeded — returns public URL path
          content:
            application/json:
              schema:
                type: object
                required: [url]
                properties:
                  url:
                    type: string
        "400":
          description: Missing file or unsupported type
        "401":
          description: Unauthenticated
        "403":
          description: Forbidden
        "413":
          description: File too large (max 10 MB)
        "500":
          description: Upload failed
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...